Skip `EPIPE` error.
There is an `EPIPE` error in Node.js 0.12 when I submit a found megabyte
buffer, but stop after a megabyte. I'm only trying to test an edge case,
the unlikely case that `getRawBody` emits an error for any reason, so
simply overwriting the limit by one byte is enough to satisfy the
requirements of the test.

Otherwise, I'd be loading up the test with socket error handling.